+++ src/engine/index.ts
@@ -135,6 +135,7 @@ export * from './Util/Toaster';
 export * from './Util/StateMachine';
 export * from './Util/Future';
 export * from './Util/Semaphore';
+export * from './Util/Coroutine';
 
 // ex.Deprecated
 // import * as deprecated from './Deprecated';